requirements.txt
# FastAPI + Uvicorn
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# HTTP client
httpx>=0.26.0

# Google Cloud Vision API
google-cloud-vision>=3.4.5

# Claude (Anthropic) via LangChain
langchain-anthropic>=0.1.11

# LangChain integrations
langchain-community>=0.1.16
langchain>=0.1.16
langchain-chroma>=0.1.4
langchain-huggingface>=0.0.5

# Vector DB + HuggingFace
chromadb>=0.4.24
huggingface-hub>=0.21.4
transformers>=4.41.0
torch>=2.2.2

# Claude-compatible .env loader
python-dotenv>=1.0.1

# SerpAPI
serpapi>=0.1.4

# Document handling
pandas>=2.2.1
python-docx>=1.1.0
pypdf>=3.17.1

# Terminal formatting
rich>=13.7.1

# Utilities
tqdm>=4.66.4
pydantic>=2.7.1
google-search-results
sentence-transformers
python-multipart
pypdf2
pdf2image
pillow
pymongo
motor
PyMuPDF
openai
reportlab

# Face Analysis Dependencies
mediapipe>=0.10.8
opencv-python>=4.8.1
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
